% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_distance_metric.R
\name{gw_distance_metric_names}
\alias{gw_distance_metric_names}
\title{List valid GW distance metric names}
\usage{
gw_distance_metric_names()
}
\value{
Character vector of valid preset names.
}
\description{
List valid GW distance metric names
}
